% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.nflseedR_simulation}
\alias{autoplot.nflseedR_simulation}
\alias{plot.nflseedR_simulation}
\title{Automatically Plot nflseedR_simulation Object}
\usage{
autoplot.nflseedR_simulation(object, type = c("wins", "rank", "points"), ...)

\method{plot}{nflseedR_simulation}(x, ..., type = c("wins", "rank", "points"), y)
}
\arguments{
\item{object}{a \code{nflseedR_simulation} object as created by \code{simulate_nfl()}}

\item{type}{one of "wins", "rank", "points"}

\item{...}{unused, required by autoplot generic}

\item{x}{A \code{nflseedR_simulation} object.}

\item{y}{Ignored, required for compatibility with the \code{plot()} generic.}
}
\value{
a ggplot object
}
\description{
Creates automatic plots for wins, ranks, or points for an \code{nflseedR_simulation} object as created by \code{simulate_nfl()}.
}
\details{
This function requires the following optional dependencies:
\itemize{
\item "forcats",
\item "ggplot2",
\item "ggdist",
\item "ggtext",
\item "ggthemes",
\item "gridtext"
}
}
\examples{
\donttest{

  simulation <- .ffs_cache("foureight_sim.rds")

  ggplot2::autoplot(simulation) # default is type = "wins"
  ggplot2::autoplot(simulation, type = "rank")
  ggplot2::autoplot(simulation, type = "points")

}

}
\seealso{
\code{vignette("Basic Simulations")} for example usage
}
