% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dictionary_game_summary}
\alias{dictionary_game_summary}
\title{Data Dictionary: Simulations | Game Summary}
\format{
An object of class \code{data.frame} with 11 rows and 2 columns.
}
\usage{
dictionary_game_summary
}
\description{
A dataframe containing the data dictionary of the simulation output table "game_summary"
}
\keyword{datasets}
