% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dictionary_games}
\alias{dictionary_games}
\title{Data Dictionary: Simulations | Games}
\format{
An object of class \code{data.frame} with 9 rows and 2 columns.
}
\usage{
dictionary_games
}
\description{
A dataframe containing the data dictionary of the simulation output table "games"
}
\keyword{datasets}
