% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dictionary_overall}
\alias{dictionary_overall}
\title{Data Dictionary: Simulations | Overall}
\format{
An object of class \code{data.frame} with 11 rows and 2 columns.
}
\usage{
dictionary_overall
}
\description{
A dataframe containing the data dictionary of the simulation output table "overall"
}
\keyword{datasets}
