% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dictionary_standings}
\alias{dictionary_standings}
\title{Data Dictionary: Simulations | Standings}
\format{
An object of class \code{data.frame} with 21 rows and 2 columns.
}
\usage{
dictionary_standings
}
\description{
A dataframe containing the data dictionary of the simulation output table "standings"
}
\seealso{
\url{https://nflseedr.com/articles/articles/nflsim2.html#simulation-output}
}
\keyword{datasets}
