% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dictionary_team_wins}
\alias{dictionary_team_wins}
\title{Data Dictionary: Simulations | Team Wins}
\format{
An object of class \code{data.frame} with 4 rows and 2 columns.
}
\usage{
dictionary_team_wins
}
\description{
A dataframe containing the data dictionary of the simulation output table "team_wins"
}
\keyword{datasets}
