% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sharpe_games.R
\name{load_sharpe_games}
\alias{load_sharpe_games}
\title{Load Lee Sharpe's Games File}
\usage{
load_sharpe_games()
}
\value{
A data frame containing the following variables for all NFL games
since 1999:
\describe{
\item{game_id}{The ID of the game as assigned by the NFL. Note that this value matches the \code{game_id} field in nflscrapR if you wish to join the data.}
\item{alt_game_id}{This is a more human-readable ID. It consists of: The season, an underscore, the two-digit week number, an underscore, the away team, an underscore, the home team.}
\item{season}{The year of the NFL season. This reperesents the whole season, so regular season games that happen in January as well as playoff games will occur in the year after this number.}
\item{game_type}{What type of game? One of the following values:
\itemize{
\item{\code{REG}}{: a regular season game}
\item{\code{WC}}{: a wildcard playoff game}
\item{\code{DIV}}{: a divisional round playoff game}
\item{\code{CON}}{: a conference championship}
\item{\code{SB}}{: a Super Bowl}
}
}
\item{week}{The week of the NFL season the game occurs in. This will be 1-17 for the regular season, 18 for wildcard playoff games, 19 for divisional playoff games, 20 for conference championships and 21 for Super Bowls.}
\item{gameday}{The date on which the game occurred.}
\item{weekday}{The day of the week on which the game occcured.}
\item{gametime}{The kickoff time of the game. This is represented in 24-hour time and the Eastern time zone, regardless of what time zone the game was being played in.}
\item{away_team}{The away team.}
\item{away_score}{The number of points the away team scored. Is \code{NA} for games which haven't yet been played.}
\item{home_team}{The home team. Note that this contains the designated home team for games which no team is playing at home such as Super Bowls or NFL International games.}
\item{home_score}{The number of points the home team scored. Is \code{NA} for games which haven't yet been played.}
\item{location}{Either \code{Home} if the home team is playing in their home stadium, or \code{Neutral} if the game is being played at a neutral location. This still shows as \code{Home} for games between the Giants and Jets even though they share the same home stadium.}
\item{result}{The number of points the home team scored minus the number of points the visiting team scored. Equals \code{h_score - v_score}. Is \code{NA} for games which haven't yet been played. Convenient for evaluating against the spread bets.}
\item{total}{The sum of each team's score in the game. Equals \code{h_score + v_score}. Is \code{NA} for games which haven't yet been played. Convenient for evaluating over/under total bets.}
\item{gsis}{The id of the game issued by the NFL Game Statistics & Information System.}
\item{pfr}{The id of the game issued by \href{https://www.pro-football-reference.com/}{Pro Football Reference}}
\item{pff}{The id of the game issued by \href{https://www.pff.com/}{Pro Football Focus}}
\item{espn}{The id of the game issued by \href{https://www.espn.com/}{ESPN}}
\item{spread_line}{The spread line for the game. A positive number means the home team was favored by that many points, a negative number means the away team was favored by that many points. This lines up with the \code{result} column.}
\item{total_line}{The total line for the game.}
\item{roof}{What was the status of the stadium's roof? Will be one of the following values:
\itemize{
\item{\code{outdoors}}{: An outdoor stadium}
\item{\code{open}}{: Stadium has a retractable roof which was open}
\item{\code{closed}}{: Stadium has a retractable roof which was closed}
\item{\code{dome}}{: An indoor stadium}
}
}
\item{surface}{What type of ground the game was played on}
\item{temp}{The temperature at the stadium (for \code{outdoors} and \code{open} only)}
\item{wind}{The speed of the wind in miles/hour (for \code{outdoors} and \code{open} only)}
\item{away_coach}{Name of the head coach of the away team}
\item{home_coach}{Name of the head coach of the home team}
\item{referee}{Name of the game's referee (head official)}
\item{stadium}{Name of the stadium}
}
}
\description{
Lee Sharpe maintains an important data set that contains
broadly used information on games in the National Football League. This
function is a convenient helper to download the file into memory without
having to remember the correct url.
}
\examples{
\donttest{
games <- load_sharpe_games()
dplyr::glimpse(games)
\dontshow{
# Close open connections for R CMD Check
future::plan("sequential")
}
}
}
