% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{nfl_standings}
\alias{nfl_standings}
\title{Compute NFL Standings}
\usage{
nfl_standings(
  games,
  ranks = c("DIV", "CONF", "DRAFT", "NONE"),
  tiebreaker_depth = c("SOS", "PRE-SOV", "RANDOM"),
  playoff_seeds = 7L,
  verbose = getOption("nflseedR.verbose", default = TRUE)
)
}
\arguments{
\item{verbose}{}
}
\description{
Compute NFL Standings
}
