% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{sims_games_example}
\alias{sims_games_example}
\title{Example Games Data used in NFL Simulations}
\format{
A data frame with 284 rows and 9 variables containing NFL schedule
information.
}
\usage{
sims_games_example
}
\description{
Example Games Data used in NFL Simulations
}
\details{
Please see \code{data-raw/sim_examples.R} for the code to create this data.
}
\examples{
str(sims_games_example)
}
\keyword{datasets}
