% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{sims_teams_example}
\alias{sims_teams_example}
\title{Example Teams Data used in NFL Simulations}
\format{
A data frame with 64 rows and 5 variables containing team name and
division information.
}
\usage{
sims_teams_example
}
\description{
Example Teams Data used in NFL Simulations
}
\details{
Please see \code{data-raw/sim_examples.R} for the code to create this data.
}
\examples{
str(sims_teams_example)
}
\keyword{datasets}
