% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_nflseedR.R
\name{summary.nflseedR_simulation}
\alias{summary.nflseedR_simulation}
\title{Compute Pretty Simulations Summary Table}
\usage{
\method{summary}{nflseedR_simulation}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments passed on to the methods (currently not used).}
}
\description{
Uses the R package gt to create a pretty html table of the
nflseedR simulation summary data frame.
}
\examples{
\donttest{
library(nflseedR)

# Activate progress updates
# progressr::handlers(global = TRUE)

# Parallel processing can be activated via the following line
# future::plan("multisession")

# Simulate the season 4 times in 2 rounds
sim <- nflseedR::simulate_nfl(
  nfl_season = 2020,
  fresh_season = TRUE,
  simulations = 4,
  sims_per_round = 2
)

# Create Summary Tables
summary(sim)
}
}
