% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{divisions}
\alias{divisions}
\title{NFL team names and the conferences and divisions they belong to}
\format{
A data frame with 36 rows and 4 variables containing NFL team level
information, including franchises in multiple cities:
\describe{
\item{team}{Team abbreviation}
\item{conf}{Conference abbreviation}
\item{division}{Division name}
\item{sdiv}{Division abbreviation}
}
This data frame is created using the \code{teams_colors_logos} data frame of the
\code{nflfastR} package. Please see \code{data-raw/divisions.R} for the code to create
this data.
}
\usage{
divisions
}
\description{
NFL team names and the conferences and divisions they belong to
}
\examples{
str(divisions)
}
\keyword{datasets}
