% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_regular_season.R
\name{nfl_regular_season}
\alias{nfl_regular_season}
\title{Compute NFL Regular Season Matchups}
\usage{
nfl_regular_season(previous_season_standings)
}
\arguments{
\item{previous_season_standings}{NFL standings of the previous season as
computed by \code{\link[=nfl_standings]{nfl_standings()}}.}
}
\value{
A data frame with one row per team/opponent matchup (this means
544 rows in a 272 games season), including the following columns:
\describe{
\item{\code{season}}{The season following the season in \code{previous_season_standings}}
\item{\code{team}}{The team abbreviation of the team we are looking at}
\item{\code{opp}}{The team abbreviation of the team's opponent}
\item{\code{opp_type}}{The opponent type. This will be on of the following.
\describe{
\item{\code{div_a}}{Divisional matchup on the road}
\item{\code{div_h}}{Divisional matchup at home}
\item{\code{intra_conf}}{Full division matchup intra conference}
\item{\code{cross_conf}}{Full division matchup cross conference}
\item{\code{div_rank}}{Same division rank matchup intra conference}
\item{\code{cross_17th}}{17th game, same division rank matchup cross conference}
}
}
\item{\code{location}}{Either \code{home} or \code{away}. It's impossible to predict neutral site games}
}
}
\description{
Use standings of a completed season to compute all matchups (including location)
of the following season based on NFL regular season schedule rules.
}
\details{
Since the 2021 season, NFL teams play a 17 game schedule. NFL operations
explains the anatomy of a schedule here
\url{https://operations.nfl.com/gameday/nfl-schedule/creating-the-nfl-schedule/}.

A team's opponents are determined based on the following rules:
\itemize{
\item 6 games against divisional opponents
\item 4 games against teams from a division within its conference
\item 4 games against teams from a division in the other conference
\item 2 games against teams from the remaining divisions in its conference that
finished the previous season on the same division rank
\item 1 game ("the 17th game") against a non-conference opponent from a division
that the team is not scheduled to play and that finished the previous season
on the same division rank
}

The host of matchups rotates regularly. Rotation has different cycles as
explained below. Here is how hosts are determined:
\itemize{
\item Each team hosts 3 division matchups
\item Intra conference division matchups rotate in a 6 year cycle
\item Cross conference division matchups rotate in a 8 year cycle
\item Intra conference divisional rank matchups rotate in a 6 year cylce
\item AFC teams host the 17th game in odd years, NFC teams host in even years
}

All rotation tables are stored inside nflseedR and used to compute the
compete regular season matchup table.
}
\examples{
# Compute all 2026 matchups using standings from the 2025 season

# First load games of the 2025 season
games <- nflreadr::load_schedules(2025)

# Second compute standings
standings <- nflseedR::nfl_standings(games, ranks = "DIV")

# Finally compute matchups
matchups <- nflseedR::nfl_regular_season(standings)

# Overview output
str(matchups)
}
