% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflseedR-package.R
\docType{package}
\name{nflseedR-package}
\alias{nflseedR}
\alias{nflseedR-package}
\title{nflseedR: Functions to Efficiently Simulate and Evaluate NFL
    Seasons}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A set of functions to simulate National Football
    League seasons including the sophisticated tie-breaking procedures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nflseedr.com}
  \item \url{https://github.com/nflverse/nflseedR}
  \item Report bugs at \url{https://github.com/nflverse/nflseedR/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

Authors:
\itemize{
  \item Lee Sharpe [copyright holder]
}

}
\keyword{internal}
