% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflseedR-package.R
\docType{package}
\name{nflseedR-package}
\alias{nflseedR}
\alias{nflseedR-package}
\title{nflseedR: Functions to Efficiently Simulate and Evaluate NFL
    Seasons}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A set of functions to simulate National Football
    League seasons including the sophisticated tiebreaking procedures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.nflseedr.com}
  \item \url{https://leesharpe.github.io/nflseedR}
}

}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com} [contributor]

Authors:
\itemize{
  \item Lee Sharpe
}

}
\keyword{internal}
